systemd check on linux before service file code

+++ .shell_aliases_functions.sh
@@ -392,6 +392,7 @@ r(){
     restartpoll(){
         src_dir="$HOME/forkedRepos/$REPO_NAME"
         test -d "$src_dir" || { echo "$src_dir does not exists." >&2 && return 1; }
+        test -d "/etc/systemd/system" || { echo "/etc/systemd/system does not exists. Is systemd installed?" >&2 && return 1; }
         git -C "$src_dir" pull
         perl -i -pe "s@pi@$USER@g" "$src_dir/poll.service"
         sudo cp "$HOME/forkedRepos/$REPO_NAME/poll.service" /etc/systemd/system
